Extract game option constants
