+++ src/structures/game_session.ts
@@ -42,12 +42,12 @@ import EliminationScoreboard from "./elimination_scoreboard";
 import TeamScoreboard from "./team_scoreboard";
 import { getRankNameByLevel } from "../commands/game_commands/profile";
 import EliminationPlayer from "./elimination_player";
-import { KmqImages } from "../constants";
+import { KmqImages, SONG_START_DELAY } from "../constants";
 import MessageContext from "./message_context";
 import KmqMember from "./kmq_member";
 import { calculateTotalRoundExp } from "../commands/game_commands/exp";
 import Player from "../structures/player";
-import Session, { SONG_START_DELAY } from "./session";
+import Session from "./session";
 import Round from "./round";
 import QueriedSong from "../interfaces/queried_song";
 import GuessResult from "../interfaces/guess_result";
